requirements.txt
tensorflow-gpu==1.11.0
absl-py>=0.5.0
numpy==1.15.1
scikit-learn==0.20.1
wget>=3.2
